feito salvar arquivo filesystem e testes

+++ dao/PostgresPostDao.py
@@ -2,6 +2,7 @@
 
 import psycopg2
 from Post import Post
+from utils import delete_arquivo
 
 
 class PostgresPostDao(object):
@@ -16,8 +17,8 @@ class PostgresPostDao(object):
         if isinstance(post, Post):
             binary = psycopg2.Binary(post.imagem_binary)
             self.query.execute(
-                "INSERT INTO post (titulo, conteudo, data, imagem) VALUES (%s, %s, %s, %s)",
-                (post.titulo, post.conteudo, post.data, binary)
+                "INSERT INTO post (titulo, conteudo, data, imagem, arquivopath) VALUES (%s, %s, %s, %s, %s)",
+                (post.titulo, post.conteudo, post.data, binary, post.arquivopath)
             )
             self.conn.commit()
             self.query.execute("SELECT MAX(id) FROM post")
@@ -31,8 +32,8 @@ class PostgresPostDao(object):
         if isinstance(post, Post):
             binary = psycopg2.Binary(post.imagem_binary)
             self.query.execute(
-                "UPDATE post SET titulo=%s, conteudo=%s, imagem=%s WHERE id=%s",
-                (post.titulo, post.conteudo, binary, post.id)
+                "UPDATE post SET titulo=%s, conteudo=%s, imagem=%s, arquivopath=%s WHERE id=%s",
+                (post.titulo, post.conteudo, binary, post.arquivopath, post.id)
             )
             consult = self.query.rowcount
             if consult:
@@ -46,7 +47,7 @@ class PostgresPostDao(object):
         self.query.execute("SELECT * FROM post")
         postsQuery = self.query.fetchall()
         for post in postsQuery:
-            posts.append(Post(post[3], post[2], data=post[1], imagem_binary=post[4], id=post[0]))
+            posts.append(Post(post[3], post[2], data=post[1], imagem_binary=post[4], arquivopath=post[5], id=post[0]))
         return posts
 
     def get_um(self, id):
@@ -56,19 +57,22 @@ class PostgresPostDao(object):
         )   
         consult = self.query.fetchone()
         if consult:
-            post = Post(consult[3], consult[2], imagem_binary=consult[4], data=consult[1], id=consult[0])
+            post = Post(consult[3], consult[2], imagem_binary=consult[4], data=consult[1], arquivopath=consult[5], id=consult[0])
             return post
         return False
 
     def delete_um(self, id):
-        self.query.execute(
-            "DELETE FROM post WHERE id=(%s)",
-            (id, )
-        )
-        consult = self.query.rowcount
-        if consult:
-            self.conn.commit()
-            return True
+        post = self.get_um(id)
+        if post:
+            self.query.execute(
+                "DELETE FROM post WHERE id=(%s)",
+                (id, )
+            )
+            consult = self.query.rowcount
+            if consult:
+                delete_arquivo(post.arquivopath)
+                self.conn.commit()
+                return True
         return False
 
     def delete_all(self):
